Shout Edit screen and Profile screen

+++ app/src/main/java/cz/davmarek/shouts/ui/screens/ShoutDetailScreen.kt
@@ -1,6 +1,7 @@
 package cz.davmarek.shouts.ui.screens
 
 import android.util.Log
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -9,16 +10,12 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Edit
-import androidx.compose.material.icons.filled.Refresh
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -28,8 +25,8 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
+import cz.davmarek.shouts.ui.components.ConfirmDialog
 import cz.davmarek.shouts.ui.components.NavigationBackButton
-import cz.davmarek.shouts.ui.theme.ShoutsTheme
 import cz.davmarek.shouts.utils.formatDateForUI
 import cz.davmarek.shouts.viewmodels.ShoutDetailViewModel
 import java.util.Date
@@ -37,9 +34,7 @@ import java.util.Date
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ShoutDetailScreen(
-    navController: NavController?,
-    viewModel: ShoutDetailViewModel,
-    shoutId: String
+    navController: NavController?, viewModel: ShoutDetailViewModel, shoutId: String
 ) {
 
     val context = LocalContext.current
@@ -50,41 +45,35 @@ fun ShoutDetailScreen(
 
     val viewState = viewModel.viewState.collectAsState()
 
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = {
-                    Text(
-                        text = "Post"
-                    )
-                },
-                navigationIcon = {
-                    NavigationBackButton(
-                        onClick = {
-                            navController?.popBackStack()
-                        }
-                    )
-                },
+    Scaffold(topBar = {
+        TopAppBar(title = {
+            Text(
+                text = "Post"
+            )
+        }, navigationIcon = {
+            NavigationBackButton(onClick = {
+                navController?.popBackStack()
+            })
+        },
 
-                actions = {
+            actions = {
 
-                    if (viewState.value.isShoutMine) {
-                        IconButton(onClick = {
-                            // TODO: Implement editing screen
-                        }) {
-                            Icon(Icons.Default.Edit, contentDescription = "Edit")
-                        }
-                        IconButton(onClick = {
-                            viewModel.openDeleteDialog()
-                        }) {
-                            Icon(Icons.Default.Delete, contentDescription = "Delete")
-                        }
+                if (viewState.value.isShoutMine) {
+                    IconButton(onClick = {
+                        navController?.navigate("ShoutEditScreen/${shoutId}")
+                    }) {
+                        Icon(Icons.Default.Edit, contentDescription = "Edit")
+                    }
+                    IconButton(onClick = {
+                        viewModel.openDeleteDialog()
+                    }) {
+                        Icon(Icons.Default.Delete, contentDescription = "Delete")
                     }
                 }
+            }
 
-            )
-        }
-    ) { innerPadding ->
+        )
+    }) { innerPadding ->
         Column(
             Modifier
                 .fillMaxSize()
@@ -93,23 +82,29 @@ fun ShoutDetailScreen(
         ) {
 
             when {
-                viewState.value.openDeleteDialog ->
-                    DeleteDialog(
-                        onDismissRequest = {
-                            viewModel.closeDeleteDialog()
-                        },
-                        onConfirmation = {
-                            viewModel.closeDeleteDialog()
-                            viewModel.deleteShout()
-                            navController?.popBackStack()
-                            Log.d("ShoutDetailScreen", "Delete confirmed")
-                        },
-                    )
+                viewState.value.openDeleteDialog -> ConfirmDialog(
+                    onDismissRequest = {
+                        viewModel.closeDeleteDialog()
+                    },
+                    onConfirmation = {
+                        viewModel.closeDeleteDialog()
+                        viewModel.deleteShout()
+                        navController?.popBackStack()
+                        Log.d("ShoutDetailScreen", "Delete confirmed")
+                    },
+                )
             }
 
             Text(
                 text = "@" + (viewState.value.shout?.user?.username ?: ""),
-                style = MaterialTheme.typography.titleLarge
+                style = MaterialTheme.typography.titleLarge,
+                modifier = Modifier.clickable(
+                    onClick = {
+                        viewState.value.shout?.userId?.let {
+                            navController?.navigate("UserDetailScreen/$it")
+                        }
+                    }
+                ),
             )
 
             Spacer(modifier = Modifier.height(8.dp))
@@ -130,46 +125,6 @@ fun ShoutDetailScreen(
     }
 }
 
-@Composable
-fun DeleteDialog(
-    onDismissRequest: () -> Unit,
-    onConfirmation: () -> Unit,
-    dialogTitle: String = "Delete Shout",
-    dialogText: String = "Are you sure you want to delete this shout?",
-) {
-    AlertDialog(
-        icon = {
-            Icon(Icons.Default.Delete, contentDescription = "Example Icon")
-        },
-        title = {
-            Text(text = dialogTitle)
-        },
-        text = {
-            Text(text = dialogText)
-        },
-        onDismissRequest = {
-            onDismissRequest()
-        },
-        confirmButton = {
-            TextButton(
-                onClick = {
-                    onConfirmation()
-                }
-            ) {
-                Text("Confirm")
-            }
-        },
-        dismissButton = {
-            TextButton(
-                onClick = {
-                    onDismissRequest()
-                }
-            ) {
-                Text("Dismiss")
-            }
-        },
-    )
-}
 
 @Preview
 @Composable
